Show courses

+++ src/components/Table.js
@@ -3,7 +3,7 @@ import './Table.css'; // Importa os estilos CSS
 import { BiTrash } from "react-icons/bi";
 import studyNestApi from '../services/apiStudyNest';
 
-export default function Table ({title, content, deleteURL, baseURL = null}) {
+export default function Table ({title, content, deleteURL = null, baseURL = null}) {
   const handleRedirect = (e, route) => {
     if(e.target.tagName === 'DIV' && baseURL)
       window.location.href = `${baseURL}/${route}`;
@@ -11,7 +11,7 @@ export default function Table ({title, content, deleteURL, baseURL = null}) {
 
   const handleDelete = (id) => {
     const token = localStorage.getItem('authToken');
-    if(token){
+    if(token && deleteURL){
 
       studyNestApi.delete(`${deleteURL}/${id}`,
         {
@@ -38,7 +38,7 @@ export default function Table ({title, content, deleteURL, baseURL = null}) {
             onClick={(e) => handleRedirect(e, id)}
             className={!baseURL ? 'cursor-default' : ''}>
               {name}
-              <BiTrash className='icon' onClick={() => handleDelete(id)} />
+              {deleteURL && <BiTrash className='icon' onClick={() => handleDelete(id)} />}
             </div>
         })
       }
